[ADD] SHLVL handling  [ADD] Startup_routine
