+++ Makefile
@@ -192,6 +192,7 @@ SRC += init_shell.c
 SRC += init_cfg.c
 SRC += destructor.c
 SRC += routine_exit.c
+SRC += startup_routine.c
 SRC += $(addprefix line_edition/,$(LESRCS))
 SRC += $(addprefix lexer/,$(LEXSRCS))
 SRC += $(addprefix parser/,$(PARSRCS))
