+++ includes/sh.h
@@ -40,11 +40,17 @@
 # define EXFD2				"Bad file descriptor\n"
 # define EXUEPTD			"Unexpected error occurred... Shall leave\nexit\n"
 
+# include "struct.h"
 # include <stdint.h>
 # include <stddef.h>
 
 typedef struct s_cfg	t_cfg;
 
+int			lexer_routine(char **line, t_lexer *lexer);
+int			parser_routine(t_lexer *lexer, t_parser *parser);
+int			line_edition_routine(char **line);
+int			eval_routine(t_parser *parser);
+int			analyzer_routine(t_cmd_table *cmd);
 void		set_var(t_cfg *shell);
 void		exit_routine(t_cfg *shell, uint8_t ret);
 void		ft_ex(char *error);
